Connect Multiple Devices
Bulk of the coding is done to support multiple devices, but I’ve only
tested with a single device. I’ll be able to test in a couple weeks
when I get a couple extra heart rate monitors.

The documentation is also fairly out of date, but should be pretty
explanatory. Basically every call needs a device address now.

And lastly, I plan to remove the example and create an example app
instead. This will make it easier to show everything that the app
supports as well as make it easier to test for myself and others.