Update plugin.xml to so can build for Windows 10 universal target
Currently trying to build for Windows 10 universal target will result in an error, because the "m2" namespace used for the DeviceCapability entries is not valid. Until this is fixed in Cordova itself, a work-around is to include a separate <config-file> element for the Windows 10 (and greater) target.

Note that this 'versions' attribute is only supported by Cordova 5.0.0 or higher. Therefore I have also updated the <engines> element to require that (note that until very recently, the 'cordova' engine did not actually specify the required Cordova version - you had to use 'cordova-plugman').