Added isConnectable to ScanParams interface
Docs mention isConnectable param as an option for startScan(), but ScanParams does not include it.

Windows 10 UWP
isConnectable - Windows 10 will, by default, show all devices ever seen by the system, even if the device is off. If you want to only devices that are on and connectable, set this to true.